[FEATURE] Disable Upload form in TCA form globaly
Currently to disable to form upload in TCA form we must use the
disable_controls key in the config array. On some TYPO3 installation
it's part of the project guidline, that the user must upload files to
fileadmin, before using them.

Disabling upload control, for each field is error prone. So a global
User TSConfig can be a good idea

Change-Id: I77786a474226ff36d2e5e24a9e9345208bca6354
Resolves: #39081
Releases: 6.0