[TASK][CLEANUP] Remove obsolete IPv6 test condition
The defined('AF_INET6') in the t3lib_div ipv6 related methods tests
for availability of ipv6 support in php. This is by default enabled
on all platforms since php 5.3.
The fallback layer in t3lib_div did not work anyway and the unit
tests failed in this case. This fallback layer is now removed.

Change-Id: Id2618c33a3ab0d5daa7faa3fd42ec72d51974485
Resolves: #39125
Releases: 6.0