[!!!][TASK] Remove old XCLASS compatibility layer
The XCLASS handling that was changed with #31893 allows to register
XCLASS'es with an entry in ext_autoload. The patch introduced a
compatibility layer to support the old registration by guessing the
needed array key in TYPO3_CONF_VARS. This turns out to not work very
reliable and could only be fixed with a massive organizational and
performance overhead.

Therefore, this patch now drops the complete compatibility layer, so an
extension must be adapted to support the new scheme. This is ok, since
extension authors maintaining extensions with XCLASS'es are aware that
those classes already break frequently in the past, and were already
advised to request hooks in the long run.

Change-Id: I95a8bf986c94c784c69d4c4cc4a9c4645cd3fcc9
Resolves: #39128
Releases: 6.0