[FEATURE] Migrate EXT: beuser to Extbase
This patchset includes a new backend user administration.

Based on extbase a model for BackendUsers is included.
All users can be listed and the listing can be filtered by
* username
* usergroup
* admin/non-admin
* active/inactive
* ever logged in.

The user listing includes the last login column also the
limit of 1.000 users was removed and pagination integrated.

Selected users can be compared in a table view.
All active backend user sessions are listed - can be terminated.

Of course the SU/change user feature was migrated.
Clickmenu was widely integrated to allow editing right away.

Change-Id: I32569c848581a49b3948e093d717dad50e973c5d
Resolves: #39141
Resolves: #5288
Resolves: #37949
Resolves: #33427
Resolves: #18602
Resolves: #15494
Releases: 6.0