[BUGFIX] CLI argument handling chokes on = use
This is a backport of the fix done in FLOW3.
The corresponding FLOW3 ticket is #30493

With this fix the last argument may use the equals sign (=)
again to separate between argument name and value.

Besides writing an additional test case which reproduces the issue,
the other test cases have been strengthened by replacing assertEquals()
by assertSame().

Change-Id: I7d7fa4e5f18a73308cc564db152fd31ab3c4e478
Fixes: #36745, #35207
Relates: #30493
Releases: 6.0, 4.7, 1.4
Reviewed-on: http://review.typo3.org/12960
Reviewed-by: Jochen Rau
Tested-by: Jochen Rau
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn