[-FEATURE] Remove restricted class prefixes
Sometimes the core enforces classes to be prefixed with special names
like 'tx_' or 'user_', especially if they are loaded as user functions.
This constraint is in the core for a long time, but was already weakened
by for example the extbase code, and was only required at some arbitrary
points. The constraint itself does not make much sense, and is now
fully removed with this patch.
As an additional advantage, this patch eases the transition to
namespaced class names.

Notes:
- Remove configuration Setting '[sys][additionalAllowedClassPrefixes]'
- Update t3lib_registry to accept every namespace at least 2 chars long
- Adapt tests
- Remove and deprecate calls to t3lib_div::hasValidClassPrefix

Change-Id: I0ba4fd41beb5f05bf87a14d735bb83922d62b247
Resolves: #39164
Releases: 6.0