[FEATURE] Let select() fetch records without translation parent
When you fetch records through the TypoScript object
"CONTENT" (or, let's say through ".select" in general),
you have an option called "languageField" in order
to fetch only records that match the current language.

What "select" does here, is that it fetches all records in
the default language ("0,-1") and then overlays them
with the records of the current language.

However, it's not possible to fetch the records that
don't have a localization parent.

The patch introduces a new option called
"includeRecordsWithoutDefaultTranslation" that also
checks for records with the languageField matching the
current language, where there is no translation parent
(l10n_parent = 0).

Change-Id: I48dc3a3b72a1dce515c5e5dd2278137888a1fe7f
Resolves: #24005
Releases: 6.2
Reviewed-on: https://review.typo3.org/13083
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Markus Klein
Tested-by: Markus Klein