[FEATURE] arrayExport() should recognize int keys
t3lib_utility_Array::arrayExport() currently exports array keys always
as strings. With the patch integer array keys are now recognized and
written as integers. Additionally, if all array keys are integers and
consecutive from zero to n, the key is left out completly. See the
unit tests for details.

This is useful for the extListArray in LocalConfiguration, which now
misses the array key and makes the array even more easy to handle.

Change-Id: I6233798a8d22257f542c575791f95226220703ca
Resolves: #39134
Releases: 6.0