[!!!][BUGFIX] be_user table password field is to short
When using the hash-algorithm "blowfish" in saltedpasswords
the hash is 60 chars long - therefore the field length has
been set to 60 characters in SQL, too.
Actually when converting old MD5 passwords to salted
copies, an additional identifer char (M or C) is added to
the hash which results in a 61 character hash.

The result is that login is not possible anymore after
converting the passwords like suggested by the reports
module - the database field has to be increased.

To be safe and because varchar only takes as much space
as really needed increase the field length to 100 chars.

Change-Id: I70646fe7939d41f1a7056376a85c506d10148480
Fixes: #39356
Releases: 4.5, 4.6, 4.7, 6.0