[BUGFIX] Admin fields not show in setup module
The admin functions are currently not show in User->setup module.
Affected is for example the 'Display debug messages in a new
browser window' checkbox, that is never show, even not for an admin.

This regression was introduced by a wrong refactoring in issue #39164,
the class prefix patch. Looking at the code before and after, the
regression is obvious.

Change-Id: Ic892c830f1ca0cfaccf3e53648d405b35d463377
Releases: 6.0
Related: #39164
Resolves: #39365