[FEATURE] Properly list multiple elements in notification mails
The current notification mailer cannot properly handle notification
mails for multiple elements. It seems that the code was never
intended to handle notification of multiple elements at once.

This patch adds a subtemplate for each changed record, all changed
records are listed in a proper list in the notification mail. The
subtemplate file is set with the config option
'stageNotificationEmail.messageMultipleElementRow'

The following markers can be used in the subtemplate:
'###ELEMENT_NUMBER###'
'###ELEMENT_TITLE###'
'###ELEMENT_KEY###'
'###ELEMENT_PATH###'
'###ELEMENT_PREVIEW###'
'###ELEMENT_SPLITTED_PREVIEW###'

The list is added to the message template with a new marker
'###CHANGED_ELEMENTS###'. The patch is fully compatible with old
style message templates as well. All old markers are now also
correctly filled with info on all records.

The subject and message templates are also improved to look better
for multiple records.

Additionally this patch adds a config option to set the crop limit
of record paths 'stageNotificationEmail.cropPathTitlesAfter'.

Change-Id: Ie20e88b34cc2bc5d6bc1bbd21642c758f03ae460
Resolves: #32543
Releases: 6.2