[BUGFIX] Delete move placeholders upon record deletion
If a record was moved in a draft workspace and subsequently
removed, then the move placeholder would remain and show up
as an actual record in the backend. To the end user it appears
as if the record was not deleted as he/she would now see the
placeholder.

Testplan:
* Create a new record "foo" in live workspace on
  a page where already a record exists
* switch to a workspace
* move the new record "foo" on the page (this will
  create a placeholder and a move placeholder in
  the database)
* switch back to the live workspace
* delete the record "foo"
* switch back to the workspace where you have moved
  the record

Current result:
* The move placeholder is shown in the page

Expected result:
* The record should be removed completely

Change-Id: I3b0593332650795539c2d7eac075515dbd981b45
Fixes: #39383
Releases: 6.2, 6.1, 4.5