[BUGFIX] Delete move placeholders upon record deletion
If a record was moved in a draft workspace and subsequently
removed, then the move placeholder would remain and show up
as an actual record in the backend. To the end user it appears
as if the record was not deleted as he/she would now see the
placeholder.

Change-Id: I3b0593332650795539c2d7eac075515dbd981b45
Fixes: #39383
Releases: 6.0, 4.7, 4.6, 4.5