[BUGFIX]  EXT:beuser must have frontend root TS template
Due to the current bootstrap structure of extbase, an extbase
based backend module must have a TypoScript frontend root template.
The patch introduces a descriptive exception message in case
this is not given.

Furthermore, the dependency injection configuration is moved
to an own file within the extension and moved to module.beuser
TS instead of global extbase scope.

Change-Id: Id621b856a3ad2297a20612e911186be473a63602
Resolves: #39602
Releases: 6.0