[BUGFIX] Increase log_data field size to prevent corrupted data
Stage change data is corrupted when comments longer than ~250 characters
are entered. The comment itself is lost aswell as the stage id.
This is fixed by choosing a more suitable (larger) field type for
sys_log.log_data.

Issue also exists in 4.7, 4.6, 4.5 but fix is probably not
backportable due to the change in the table definition.

Change-Id: I104bb2dbed75ebca6c063334ada543ef544f7b10
Fixes: #38403
Releases: 6.0