[BUGFIX] autoloader should ignore non existing extensions
When building the core registry a call to t3lib_extMgm::extPath() is done
for every extension of the LocalConfiguration/localconf.php
Sometimes the extension is not installed anymore which was never a problem
in earlier versions.

The change catches the exception and ignores this extension when building
the registry.

Change-Id: Ife3173ad95a0359b2bd7fcd09cfde1e6a6a3914f
Resolves: #39649
Releases: 6.0, 4.7, 4.6