[BUGFIX] autoloader should ignore non existing extensions
When building the core registry a call to t3lib_extMgm::extPath() is done
for every extension of the LocalConfiguration/localconf.php
Sometimes the extension is not installed anymore which was never a problem
in earlier versions.

The change catches the exception and ignores this extension when building
the registry.

Change-Id: I79e34758617675957c732bbd533302adbb2edab0
Resolves: #39649
Releases: 6.0, 4.7, 4.6