[BUGFIX] Avoid logout when changing the encryption key
Due to the change to hmac to generate the hash for 
the session save path in #39653 one is now logged out 
of the install tool when changing the encryption key.

Change the encryption key at runtime and ensure that the 
new session directory is created to avoid the logout.

Change-Id: Ia4625105b5c26fb7bfafd92f51d22a4466f5ca34
Fixes: #39688
Relates: #39653
Releases: 6.0
Reviewed-on: http://review.typo3.org/13569
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel