[BUGFIX] Sprite manager cache race condition
If the sprite manager is initialized in the frontend with empty caches,
the cache content will not be rendered correctly due to missing
information in frontend scope.
Solution is to not initialize the sprite manager in frontend and cli
context at all, they are not needed anyway. feedit works as before.

Change-Id: I5949f10d2383f4c259e0d86efe85960589ccf1a1
Resolves: #39706
Releases: 6.0
Reviewed-on: http://review.typo3.org/13592
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn