[FEATURE] Add new extension manager to the core
For 6.0 a new extension manager based on
extbase was developed. Goal of this extension
manager is not to be the most feature rich
but easy to use extension managers.

Therefore the whole extension manager was
restructured and some features where removed:
* language handling -> will be an own extension
* file editing -> can be done via other extensions
* upload extension -> will be integrated into extdeveval

This patch adds the base extension manager.
Styling and JS fine tuning will be done afterwards.

Please test the given functionality carefully and
report as many bugs as you can find to the project
at forge (TYPO3 6.0 > Extension Manager).

Change-Id: I28ef14401f40e239e5ea235af2be3e431fb8789d
Resolves: #39726
Releases: 6.0