[BUGFIX] Tx_Fluid_ViewHelpers_Be_Buttons_IconViewHelper broken
It seems as the sprite keys in
$GLOBALS['TBE_STYLES']['spriteIconApi']['iconsAvailable']
have changed with TYPO3 4.6 and this change was not reflected in
the experimental backend IconViewHelper, so the old key are not
longer valid and new ones cant be used because of the allowedIconArray.

I removed the "securityCheck" and the allowedIconArray, because the validity
of the icon name is checked again in t3lib_iconWorks::getSpriteIcon
against the iconAvailable array.

Change-Id: I72116e82a4ab1c703816dd6fb79e547cc1e6a9ed
Fixes: #31320
Releases: 1.4, 4.7
Reviewed-on: https://review.typo3.org/13616
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs