[BUGFIX] StorageRepository should ignore not found driver class
If a driver doesn't exist anymore, the storage repository shouldn't load
it.

Most drivers will be implemented by extensions but if this extension is
deactived, the StorageRepository still tries to load this driver which is
not possible anymore and this ends in an uncaught exception in BE.
Therefore it should be checked if an implementation is available.

Change-Id: Ib71bef82a53096746cff4bd9577d6adbeae90ac5
Fixes: #39789
Releases: 6.0