[BUGFIX] Fix failing t3lib_log tests
Some tests for t3lib_log are failing in some cases, because
the fixture classes are not available. They are available
when executing the tests with the phpunit backend module
because the files are named *_test.php, recognized as
test files and thus required.

To avoid this confusion, the fixture classes are now moved
to a fixture directory and renamed so that they not end with *_test.php
To be consistent the class names are also renamed.


Change-Id: Ia5efce2909111b79ed6c836c4c704a78faacdc65
Fixes: #39885
Releases: 6.0