[FEATURE] Report status check for file and folder create mask
In sane server setups, it is usually not a good idea to configure TYPO3
to create files and folders with writable bit for 'others'. The
introduction package actually sets fileCreateMask and folderCreateMask
to 666 and 777, but this is to ease the installation process and make
the introduction package work in curious setups as well without problems.

Therefor we now add a warning to the reports module instead, if the
write bit for others is set, so an administrator is hinted on the
possible security impact, while the installation process is still smooth.

Change-Id: Iae75a9f9492d8b784a3e1ea2c754a14abbc58f3e
Releases: 6.0
Resolves: #39912