[BUGFIX] Error occurred while trying to call updateExtensionListFromTerAction
The argument forceUpdateCheck is optional and is just set if
a user actively clicks on the update link. As it is set to
required in the code the automatic update fails as it does not
set the argument.

Releases: 6.The argument forceUpdateCheck is optional and is just set if
a user actively clicks on the update link. As it is set to
required in the code the automatic update fails as it does not
set the argument.

Change-Id: I27be68656faa4ac0c40bdc395915a9f6ecd7744f
Fixes: #39946
Releases: 6.00