[FEATURE] Add translation handling module
As the new extension manager for 6.0 does not handle translation
handling a module is needed to select and update translations of
extensions.

This change to the sysext lang extensions ands a module with
which it is possible to both change select available languages
as well update the translation files for extensions.

Change-Id: I2a4f53fa01a1ba157a2ef8093130f80e06b74b72
Fixes: #39963
Releases: 6.0