[FEATURE] Add translation handling module
As the new extension manager for 6.0 does not handle translation
handling a module is needed to select and update translations of
extensions.

This change to the sysext lang extensions ands a module with
which it is possible to both change select available languages
as well update the translation files for extensions.

Fixes: #39963
Releases: 6.0

Change-Id: I416cf78d8e1d5a23043a6d3c45b149b8bee68f58