[BUGFIX] Refine page module styling
After merging the standard layout based on backend layouts the look
of the page module has changed so that a 2px border is used on the
columns that are only as long as the content elements in the column.

This changes the page module to look more like the previous version
using a boder that runs down the full length of the longest column
on all columns.

Change-Id: I6744511771744a23ad35a2d7d6377fd8ff6f14d4
Fixes: #39971
Releases: 6.0