[BUGFIX] Fix cropping of transparent gifs with im6.
The usage of the ! operator to force cropping empty image
canvas is inverted between ImageMagick v4 and v6 while
GraphicsMagick accepts both syntaxes. Because im 4 is
deprecated, we can always use im6 syntax.

There is a related fix for the 4.5 branch on gerrit
that takes im4 into account (14132).

Related: #19045
Releases: 6.0, 4.7, 4.6
Change-Id: Iee1260a94f2686a512ce1b1df65c384319a4c4de