[BUGFIX] Test dependency to autoloader tests
The new class_alias() mapping in the autoloader makes the autoloader
behave weird if it is unregistered and registered multiple times during
one call. This is the case in unit tests and can lead to fatals in other
tests. For now, it is checked if the autoloader already loaded the class
alias map to no redo this and the internal mapping is not overwritten.

Change-Id: I18d501dd803ae01183bfb7eeb37a819935c99681
Resolves: #40147
Related: #40095
Releases: 6.0
Reviewed-on: http://review.typo3.org/14054
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn