[BUGFIX] drop declare(encoding=) when merging extension config files
Encoding of PHP files may be specified with a
> declare(encoding = 'UTF-8');
line at the top of the file.
This line is only allowed once in a file, so when merging
ext_tables.php and ext_localconf.php files into a single file,
we need to remove all the encoding declaration lines.

Change-Id: I6f55b13fc83aa928856aace912f37430d7ae3549
Fixes: #40069
Releases: 4.6