[TASK] Namespace and move t3lib unit tests - 2
This patch moves most unit tests of the tests/Unit/t3lib/ main
directory. They follow the new class names of the classes under
test and are namespaced. Covered are files that belong to the
backend and core extension.
The travis test configuration is adapted to execute tests
located in those extension folders.

Change-Id: I9b6da42a4c5f9a6bfcf9218ecc76e420b340ed6c
Resolves: #40353
Related: #40095
Releases: 6.0