[TASK] Extensionmanager: Remove persistence mapping
To make the new extensionmanager more standalone and run with less
dependencies, the table and field persistence mapping should be
removed. Therefore, table sys_ter is renamed to the standard extabe
table naming scheme following the model name "repository". Furthermore,
the fields are normalized to be in line with the exbase model property
names.

This patch is needed for the installation process of introduction and
government package.

Change-Id: I586572c06b07e1f78df789aa659496b32250df14
Resolves: #40394
Releases: 6.0