[BUGFIX] Files contents cannot be changed
Setting file contents failed for two reasons:

a) the permissions were incorrectly read from TSconfig and

b) we used a new "update" permission that was not present in the default
configuration, so without manually setting it, a file could never be
updated, even by admins.

This commit changes the "update" permission back to "edit", which was
also used in the pre-FAL era.

Change-Id: I408a19257a50868f7978ffac32eefbd410117e2d
Resolves: #37812
Resolves: #39860
Resolves: #39949
Releases: 6.0