[BUGFIX] Changing file title changes file name
The TCA configuration of sys_file allows direct editing of the file
name. As t3lib_TCEmain does not recognize this change as a
filesystem-related change on saving, the file is not renamed on disk.
Instead, the index record is changed, so the file is not recognized
anymore, leading to a reindexing. Because of this, all references to the
old record seem to break.

This commit makes the name field read-only and adds the title field
(which is not present in the TCA at all currently).

Change-Id: Ice110fa7ea0cc335f8b9fa8a57ecc9fe7527de76
Resolves: #39884
Resolves: #36795
Resolves: #39859
Releases: 6.0