[TASK] Stabilize some tests if run as root
GeneralUtility::fixPermissions has a bug if createGroup is set to the id
of the root group (0), and does not set the correct group. This is fixed
with a more sophisticated test in fixPermissions to allow 0 as
valid group. Default for createGroup is empty string, so it is now
checked with isset and strlen > 0 if chgrp() should be applied.

Additionally, the FAL test permissionsAreCorrectlyRetrieved is splitted
to four tests, and two of them are skipped if the test is run on linux
as root.

Change-Id: Icd3bb5808fef195da8394a8fd9c2797f9f01e9aa
Resolves: #40523
Releases: 6.0