[BUGFIX] LogLevelTest dataProvider not used
isValidLevelThrowsExceptionOnInvalidLevelIfAskedToDoSo() test throwed a
warning that required input parameter $inputValue was not used. This was
because of a typo @dataprovider -> @dataProvider, so the data provider
was not used at all. Fixing this, one of the tests with the
object \stdClass failed.
This turned out to be an error in the thrown exception in validateLevel
 because the exception is not thrown, if the exception message is
compiled from a string and an object. Therefore, the parameter $level
is now left out in method validateLevel().
As additional cleanup, the MathUtilityTest is moved from t3lib to core
extension, and the rootline test and fixture names are adapted.

Change-Id: If5ec0505abc224ad59cd5dc6cb4b679fc4bc46ba
Fixes: #40540
Releases: 6.0