[BUGFIX] Make FAL Folder not found handling more user friendly
If an user moves the folder he currently views in the file
list module, after reload he sees only a red FlashMessage
error because the viewed folder is gone.

As the JavaScript cannot determine what might be the
identifier of the moved folder (after moving) it cannot
redirect there. Thus the file module needs to react more
user friendly in such a situation.

As a solution the module tries parent folders in, too.

Change-Id: Id9e8d988b159c3f0a21b41c4d1782101146ad23c
Fixes: #40547
Releases: 6.0