[BUGFIX] ProcessedFiles do not use Original File if unchanged
The ProcessedFile used for rendering of frontend images
should use the original file, if the file did not change
after processing. Due to a small bug in persisting the
processedFile record, this does not happen.

As a result images in frontend are not displayed anymore,
if a file is moved in filelist, and referenced as a non-
resized image.

Change-Id: I186a46f9923dacd98be655d72be2bd89b43866c2
Fixes: #39904
Releases: 6.0