[BUGFIX] Editing *_users with saltedpasswords active fails
Since the namespace changes, the custom form-field-
evaluators for passwords shipped by saltedpasswords are
registered using the namespaced classes.

As the registered class name is used by tceForms to build
an internal JS function, the JS is broken due to the fact
that backslashes are not allowed in JS function names.

This patch reverts back to "classic" classnames as eval
identifier.

Change-Id: Ica52c01456be7a8ee2b7988cc9773586852621a6
Fixes: #40561
Releases: 6.0