[BUGFIX] Backend can't deal with fal references in translated records
The Preview-Functionality implementing showing the "images"
of a tt_content record in for example list and page module
is not able to deal with FAL files if the record to for
which the preview should be rendered is not the original
language.

Enable the preview to resolve the relations correctly.

Change-Id: I0754af0d4c84c827be2019362aeda09770f1873b
Fixes: #40610
Releases: 6.0