[BUGFIX] L10n PHP parser stores NULL for unknown key
The PHP parser caches NULL instead of an array when the Backend
language is not available for a given extension using PHP array
localization files (e.g., tscobj with German "de").

This leads to PHP warning when reading the labels as NULL cannot
be used in a foreach statement.

Change-Id: I1678674ecfba71e04c9c13aed44dd7f0aa574cbc
Fixes: #40643
Releases: 6.0, 4.7, 4.6
Reviewed-on: http://review.typo3.org/14360
Reviewed-by: Philipp Gampe
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers