[TASK] CSS StyledContent wraps classes spaceBefore/After twice
Since all top level TypoScript cObject are stdWrap enabled,
setting spaceBefore and spaceAfter fails as the outputtet
classnames are wrapped twice: once within the userFunc and
a second time by the general stdWrap execution.

As the stdWrap ist only meant to act as an configuration
for the userFunc, change the name accordingly, that it is
not processed by the default stdWrap anymore.

Change-Id: Id6f91429ef67351f922f261da0aeefeddc0f81b4
Fixes: #40642
Releases: 6.0