[BUGFIX] Fatal errors for interfaces with PHP version < 5.3.7
This bugfix rewrites function type hints to match the new ones as long as
PHP versions of OS Venders are not up to date.
Fatal errors are thrown when using an old interface with the new
TYPO3 6.0 beta 1. This only happens with PHP versions below 5.3.7 as
there are some type hinting issues.

Change-Id: I87da4bb2ef0420a94d1d155edde500c1aa96894a
Fixes: #40653
Related: #40095
Releases: 6.0