[BUGFIX] Many severe bugs in ProcessedFile implementation
ProcessedFiles are used to store processed versions of files, to prevent

The current implementation causes several problems. Files used in
tt_content and pages get reindexed for/at storage 0, as they are hand
over just as file name.

As a result the processed images resided in typo3temp instead of the
configured processing folder.

A further result was that changes to the actual file in a "real" storage
cannot be tracked. Furthermore the detection if a file needs to be
reprocessed did not work as expected.

After moving files, the frontend resulted in broken images, if they had
not been resized. After deleting the table of processed_files but not
deleting the generated images, exceptions have been thrown. And
additionally deleting the processed files without deleting the entries
in the database caused blank images in the frontend.

As some problems have been conceptional a major refactoring has been done
to the ProcessedFile logic.

The existing public API has not been changed, but has been extended to
better separate the concerns.

Change-Id: I186a46f9923dacd98be655d72be2bd89b43866c2
Fixes: #39904, #40033, #40669
Releases: 6.0