[BUGFIX] FAL LocalDriverTests throws warnings
finfo() in FAL LocalDriver can not deal with vfs streams and throws
warnings. It is used to fetch the mime type of files.
The patch moves the finfo call to an own method and mocks the methods
in all tests that triggered the warnings.
Additionally, the statReturnsCorrectFileInfo() test is splitted into
four tests and renamed to getLowLevelFileInfoReturnsCorrectFile*. An
additional test is added to test for correct retrieval of mime type
with a prepared dummy file.

Change-Id: Ib3dcd0998dee3d2fa7e0e60eabcf783b712c5fdd
Resolves: #40765
Releases: 6.0