[BUGFIX] Warnings in StringUtility
StringUtility::isLastPartOfString does not sanitize the input arguments
enough and throws warnings if calling strlen() for example with arrays.
The patch changes the (new 6.0) method to sanitize the arguments and now
throws Exceptions if they can not be interpreted as string.
The tests are adapted to test for the exception and adds a test with
an object to the data provider.

Change-Id: I497a1478a7804c4bbeddc2d1e340e29d73efa200
Resolves: #40767
Releases: 6.0