[TASK] Add test integrity test
In the past it happened quite often that some test did not take care
of proper backing up globals or other environment settings. This
usually leads to failing tests if they are run afterwards. Those
side effects are often hard to track down and take quite some time
if searching for the one test that breaks the environment.

Therefor, a new "Integrity" test is added to the system that is
executed as the very last test in the list of test suites. It checks
that important things that were messed up in the past are still working.

Currently it checks that the registered singleton instances are still
there, that $GLOBALS['TYPO3_DB'] still has a valid resource link to
the database, and $GLOBALS['typo3CacheManager'] is still a valid
class instance.

If one of this test fails after any merge (especially combined with
travis), it is now obvious that this merge broke the environment, so
the evil test can be found much quicker and the evil merge can be
reverted directly.

Change-Id: I0dc733081ea9ad0c050a7dbcf4d95dd8f819f1bf
Resolves: #40832
Releases: 6.0
Reviewed-on: http://review.typo3.org/14524
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn