[BUGFIX] Configuration of an extension fails with fatal error
The extensionmanager tries to merge the default configuration
with the already saved one. If no configuration has been set yet,
the result of the unserialize operation is a boolean false instead
of an array and that is causing a fatal error when it tries to
merge both configurations with the ArrayUtility::flatten method
This patch adds a check that sets the current configuration
to an empty array if none exists.

Change-Id: Ibbc92e580e94e03452e7a40800374f086488a14b
Resolves: #40864
Releases: 6.0