[FEATURE] TS config to not generate body tag automatically
When using TemplaVoila for page generation, you sometimes need to be able to set classes directly on the body tag from within TemplaVoila. Thus TV needs to generate the page's <body> tag.

Unfortunately, TYPO3 itself generates the <body> tag, too - so we have 2 body tags in the HTML document. To solve this problem, TYPO3 needs a way to not generate a <body> tag.

The attached patch introduces the config.disableBodyTag TypoScript option. When it is set, TYPO3's page generator does not generate a <body> tag at all.

Change-Id: Ib0b93b18645cb94cb95130634cc21dfb3e5f685b
Fixes: #40986