[FEATURE] Add a TS option to disable body tag generation
When using TemplaVoila for page generation, you sometimes need to be
able to set classes directly on the body tag.

Unfortunately, TYPO3 itself generates the <body> tag, too - so we have 2
body tags in the HTML document. To solve this problem, TYPO3 needs a way
to not generate a <body> tag.

Therefore introduce the config.disableBodyTag TypoScript option. When it
is set, TYPO3's page generator does not generate a <body> tag at all.

Change-Id: Ifcbe67134a910cdf26075af45fbd2b2c0efe203b
Resolves: #40986
Releases: 6.1
Reviewed-on: https://review.typo3.org/14685
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov