[TASK] Provide a file for IDEs to recognize old class names
As of now an IDE cannot recognize old class names as the
class alias is set up dynamically.

Provide a file with class definitions of old class names
which extend the new one. This file is never included
during a request but just there to add IDE support for
the old class names.

The file is auto generated on the basis of the class mapping.

Change-Id: I5ff4a1cd6a678121621ad027fb6fd63e259e02bd
Resolves: #41128
Releases: 6.0